retructure the server code (spoiler: code is still buggy)
CURR BUG: for some reason the first player's thread starts before the
the second player even started his client. I can't seem to find a logical reason
how this happens, since the condition for starting both threads is that
there are *two* threads in the list. I am doing some reboots to make
sure this is an actual logical error and not tired hardware, then
tomorrow  I'll
try to debug this. Feel free to join the debugging effort :)